osucass ItemRenderer modifications (#32)
* Removed tools menu iframe xy coordinate adjustment

The code that was supposed to fix the xy coordinates for the item tools
menu when the blackbox is part of an iframe was moving the menu to the
incorrect location on the screen. Removing the xy adjustment causes the
tools menu to appear on the correct part of the screen.

The code that is supposed to adjust the x and y coordinates for the item
tools menu when it

* overload getRenderableDocument to load accomms

getRenderableDocument hardcoded the language as english and did not pass
in the accommodations. This prevented the accommodations that need to be
set in the page from beign loaded. I overloaded the method so that
accommodations can be passed in and loaded into the document.

* Moved scripts from  Iris to Blackbox

In order to load some accommodations in Iris the Scripts in blackbox
needed to be extended. In order to follow the same convention as the
Student module those scripts were moved to Blackbox.

* Added missing semicolon.

The missing emicolon caused errors when reloading items.

* Encode spaces in the base URI

The ConfigBuilder would break and throw a URISyntaxException
if the project is being run from a directory with spaces in
the path. Encoding any spaces in the path to %20 solves
this issue.

* calculator uses absolute path

The calculator now uses the absoute path for the calculator page so that
it can find it correctly.

* Added iris dictionary files to blackbox

Added the required JavaScript files for the dictionary accommodation to
function in Iris. The scripts are loaded in scripts_iris.xml.

* Use base url instead of window location

The base url is constant and always worked. Using location.href breaks the
dictionary if someone clicks on the toolbar then loads another item.

* Add missing class to note container div

Without the 'bd' class the text area in the notes container will not have the
CSS applied to it that scales the text area to the correct width and height for
the containing frame.

* Prevent multiple Illustration resize handles

* Add webpack script for combining js files

* added shared webpack

* Add script bundles

* updated scripts_iris.xml

* updated project css to point

* removed styles

* updated bundle script to exclude styles

* updated bundle scripts to fix lang code

* Made english glossary appear first, removed logic for illustration
glossary appearing first.

* Updated bundled and minified scripts

* Removed webpack bundle scripts

* Moved files added by osu-cass to renamed directory

* Removed script bundles

We bundled the scripts used by Iris. These have been moved to Iris.